requirements.txt
google-cloud
google-cloud-storage
google-cloud-pubsub
google-api-python-client
google-api-core
firebase-admin
numpy
pillow